Splitting up GameArea
